refactor(cli): extract CrossProjectProductReferenceGenerator into its own file
The reference-proxy helper was a file-private struct at the bottom of
CrossProjectTargetDependencyGenerator.swift, which mixed two concerns in
one file: collecting the cross-project edges to wire, and building the
product references those edges need.

Moving it out required widening it from private to internal, since the
two types no longer share a file. It stays module-internal, so this adds
nothing to the module's public surface.

The rationale for preferring a proxy over a plain BUILT_PRODUCTS_DIR
path only existed in the other file's header, so the type now carries
its own doc comment rather than losing that context in the move. Imports
are narrowed to what the extracted code uses: TuistCore and XcodeGraph
were needed only by the code left behind.

No behaviour change. GraphTraverserTests (166) and
CrossProjectTargetDependencyGeneratorTests (10) pass, the workspace
builds, and the linter is clean on both files.
